rename clevar.yaml -> clevar.clv_yaml
