+++ bin/clevar_artificial_footprint
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import argparse
-from clevar import yaml
+from clevar import clv_yaml
 
 if __name__=="__main__":
     # Get command line arguments
@@ -11,5 +11,5 @@ if __name__=="__main__":
     parser.add_argument('-c','--case', dest='case', default='both', help='Run for each catalog? Options: 1, 2, both')
     args = parser.parse_args()
     # Run function
-    yaml.artificial_footprint(args.config_file, args.overwrite_config,
+    clv_yaml.artificial_footprint(args.config_file, args.overwrite_config,
                          args.overwrite_files, args.case)
