+++ bin/clevar_match
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import argparse
-from clevar import yaml
+from clevar import clv_yaml
 
 if __name__=="__main__":
     # Get command line arguments
@@ -10,4 +10,4 @@ if __name__=="__main__":
     parser.add_argument('-om','--overwrite-files', dest='overwrite_files', default=False, action='store_true', help='Overwrites matching output files')
     args = parser.parse_args()
     # Run function
-    yaml.match(args.config_file, args.overwrite_config, args.overwrite_files)
+    clv_yaml.match(args.config_file, args.overwrite_config, args.overwrite_files)
