+++ docs/doc-config.ini
@@ -7,7 +7,7 @@ geometry
 match
 match_metrics
 utils
-yaml
+clv_yaml
 
 DEMO
 ../examples/catalogs.ipynb
